+++ src/connector.rs
@@ -41,6 +41,8 @@ use crate::{
     sync::{
         sync,
         State,
+        CONVEX_CURSOR_TABLE,
+        CONVEX_CURSOR_TABLE_COLUMN,
     },
 };
 
@@ -62,7 +64,7 @@ impl ConvexConnector {
 
         let columns = source.get_columns().await?;
 
-        let tables = TableList {
+        let mut tables = TableList {
             tables: columns
                 .into_iter()
                 .map(|(table_name, column_names)| Table {
@@ -89,6 +91,15 @@ impl ConvexConnector {
                 })
                 .collect(),
         };
+        tables.tables.push(Table {
+            name: CONVEX_CURSOR_TABLE.to_string(),
+            columns: vec![Column {
+                name: CONVEX_CURSOR_TABLE_COLUMN.to_string(),
+                r#type: DataType::Long as i32,
+                primary_key: true,
+                decimal: None,
+            }],
+        });
 
         // Here, `WithoutSchema` means that there is no hierarchical level above tables,
         // not that the data is unstructured. Fivetran uses the same meaning of “schema”