+++ src/sync.rs
@@ -6,6 +6,7 @@ use futures::{
     StreamExt,
 };
 use futures_async_stream::try_stream;
+use maplit::hashmap;
 use serde::{
     Deserialize,
     Serialize,
@@ -37,6 +38,11 @@ use crate::{
 
 /// The value currently used for the `version` field of [`State`].
 const CURSOR_VERSION: i64 = 1;
+
+/// The table in which the connector syncs the end-cursor of the sync.
+pub const CONVEX_CURSOR_TABLE: &str = "_convex_cursor";
+/// The column name to the CONVEX_CURSOR_TABLE table
+pub const CONVEX_CURSOR_TABLE_COLUMN: &str = "cursor";
 
 /// Stores the current synchronization state of a destination. A state will be
 /// send (as JSON) to Fivetran every time we perform a checkpoint, and will be
@@ -251,11 +257,13 @@ async fn initial_sync(
     snapshot: Option<i64>,
     cursor: Option<ListSnapshotCursor>,
 ) {
-    yield UpdateMessage::Log(
-        LogLevel::Info,
-        format!("Starting an initial sync from {source}"),
-    );
-    log(&format!("Starting an initial sync from {source}"));
+    let log_msg = if let Some(snapshot) = snapshot {
+        format!("Resuming an initial sync from {source} at {snapshot}")
+    } else {
+        format!("Starting an initial sync from {source}")
+    };
+    log(&log_msg);
+    yield UpdateMessage::Log(LogLevel::Info, log_msg);
 
     let mut snapshot: Option<i64> = snapshot;
     let mut cursor: Option<ListSnapshotCursor> = cursor;
@@ -283,12 +291,22 @@ async fn initial_sync(
         cursor = res.cursor.map(ListSnapshotCursor::from);
     }
 
-    yield UpdateMessage::Checkpoint(State::create(Checkpoint::DeltaUpdates {
-        cursor: DocumentDeltasCursor::from(snapshot.context("Missing snapshot from response")?),
-    }));
+    let cursor = DocumentDeltasCursor::from(snapshot.context("Missing snapshot from response")?);
+    yield UpdateMessage::Update {
+        schema_name: None,
+        table_name: CONVEX_CURSOR_TABLE.to_string(),
+        op_type: OpType::Upsert,
+        row: hashmap! {
+            CONVEX_CURSOR_TABLE_COLUMN.to_string() => FivetranValue::Long(cursor.0),
+        },
+    };
+
+    yield UpdateMessage::Checkpoint(State::create(Checkpoint::DeltaUpdates { cursor }));
 
     yield UpdateMessage::Log(LogLevel::Info, "Initial sync successful".to_string());
-    log(&format!("Initial sync from {source} successful."));
+    log(&format!(
+        "Initial sync from {source} successful at cursor {cursor}."
+    ));
 }
 
 /// Synchronizes the changes that happened after an initial synchronization or
@@ -297,9 +315,9 @@ async fn initial_sync(
 async fn delta_sync(source: impl Source, cursor: DocumentDeltasCursor) {
     yield UpdateMessage::Log(
         LogLevel::Info,
-        format!("Starting to apply changes from {source}"),
+        format!("Starting to apply changes from {source} starting at {cursor}"),
     );
-    log(&format!("Delta sync from {source}."));
+    log(&format!("Delta sync from {source} starting at {cursor}."));
 
     let mut cursor = cursor;
     let mut has_more = true;
@@ -321,6 +339,15 @@ async fn delta_sync(source: impl Source, cursor: DocumentDeltasCursor) {
 
         cursor = DocumentDeltasCursor::from(response.cursor);
         has_more = response.has_more;
+
+        yield UpdateMessage::Update {
+            schema_name: None,
+            table_name: CONVEX_CURSOR_TABLE.to_string(),
+            op_type: OpType::Upsert,
+            row: hashmap! {
+                CONVEX_CURSOR_TABLE_COLUMN.to_string() => FivetranValue::Long(cursor.0),
+            },
+        };
 
         // It is safe to take a snapshot here, because document_deltas
         // guarantees that the state given by one call is consistent.